Revert "uninative: Don't use single sstate for pseudo-native"
Issue: LINCD-4558

This reverts commit 3ecf5d9692fec97b37af6a4e6c747a4e2c2ca292.

The above commit causes a regression, making eSDK not able to be
installed on a different host. For example, build eSDK on ubuntu18
and install the eSDK on ubunt20, then we'll get an error at installation
time. This is because the pseudo-native's sstate cache could not be shared
among different distros.

This is a serious regression, as if eSDK could not be installed on different
distro, then it can barely be called 'SDK'. Upstream is still discussing
this problem, yet not come up with a solution.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20210310

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>