libstd-rs: Build in the release mode to avoid error
Issue: LINCD-7172

When mozjs in firefox is built with DEBUG_BUILD = "1"
in local.conf, it will fail with the following error:
   rustc-1.56.0-src/vendor/compiler_builtins/src/int/specialized_div_rem
/asymmetric.rs:57:
   more undefined references to `core::panicking::panic' follow

Switch to building libstd-rs in "release" mode as that works
around the issue and builds mozjs successfully. This is a
work-around that can be used until the root cause of the debug
build error is resolved.

(LOCAL REV: NOT UPSTREAM) -- to be sent to oe-core

Signed-off-by: naveen gowda <naveen.gowda@windriver.com>
Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>