image.bbclass: disable two systemd networkd related units
Issue: LIN1022-909 LIN1022-913

We need to disable systemd-networkd-wait-online.service
so that it does not give us 'timeout' error when
systemd-networkd does not manage anything.

We also need to disable systemd-network-generator.service
so that it does not generate networkd configuration files,
especially in case of 'ip=dhcp', which results in a general
network interface matching configuration file to be generated
to mess things up.

In fact, in presets/90-systemd.preset in systemd source
codes, these two services are disabled. However, the oe-core
specific native implementation of systemctl, only supports
enabling units. Ideally, this systemctl script should be
improved to cover the 'disable' case. But for now, let's
just disable what we need to disable.

(LOCAL REV: NOT UPSTREAM) -- WRLinux Specific Patch

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Mingli Yu <mingli.yu@windriver.com>