systemd: Set RebootWatchdogSec to 60s as watchdog
Issue: LINCD-9346

The systemd-shutdown sets watchdog timeout to 10m (600 seconds) which is too
large, and caused errors when reboot on boars such as rpi4:

systemd-shutdown[1]: Failed to set timeout to 10min: Invalid argument

The watchog's default value is 60s, so set RebootWatchdogSec to 60s to fix the
errors. And <machin.conf> can set WATCHDOG_TIMEOUT when needed, for example,
the max timeout of rpi4 is 15 seconds.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20220527

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>