rng-tools: Replace obsolete "wants systemd-udev-settle"
Issue: LIN1022-1317

The systemd-udev-settle service is listed as obsolete and does in
some cases result in a significant boot time.  Replace that with a
specific wait for the random source hardware device to be loaded.

Before this change:

   Startup finished in 4.025s (kernel) + 2min 13.236s (userspace) = 2min 17.262s
   multi-user.target reached after 2min 8.153s in userspace

After this change:

   Startup finished in 3.956s (kernel) + 13.294s (userspace) = 17.250s
   multi-user.target reached after 9.522s in userspace

(master rev: d43e283d749ddd8fd6722411d051f0e505054c6f)

Signed-off-by: Drew Moseley <drew@moseleynet.net>
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Mingli Yu <mingli.yu@windriver.com>