openssl: export OPENSSL_MODULES in SDK
Issue: LIN1022-1895

In current SDK, when running the following command in python
shell, we get an error.

$ python3
>>> from cryptography.hazmat.backends import openssl

The error message is as below:

  cryptography.exceptions.InternalError: Unknown OpenSSL error.

We could set OPENSSL_MODULES explicitly in nativesdk-openssl package
so that when SDK is set up, it's in environment and we can
get rid of the above error.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20220927

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>