libpam: fix CVE-2022-28321
Issue: LIN1022-1815

The Linux-PAM package before 1.5.2-6.1 for openSUSE Tumbleweed allows
authentication bypass for SSH logins. The pam_access.so module doesn't
correctly restrict login if a user tries to connect from an IP address
that is not resolvable via DNS. In such conditions, a user with denied
access to a machine can still get access. NOTE: the relevance of this
issue is largely limited to openSUSE Tumbleweed and openSUSE Factory;
it does not affect Linux-PAM upstream.

References:
https://nvd.nist.gov/vuln/detail/CVE-2022-28321

Upstream patches:
https://github.com/linux-pam/linux-pam/commit/08992030c56c940c0707ccbc442b1c325aa01e6d
https://github.com/linux-pam/linux-pam/commit/23393bef92c1e768eda329813d7af55481c6ca9f

(LOCAL REV: NOT UPSTREAM) -- send to oe-core(kirkstone) on 20221028

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>
Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>