ppp: Security fix CVE-2022-4603
Issue: LIN1022-2709

CVE-2022-4603:
A vulnerability classified as problematic has been found in ppp.
Affected is the function dumpppp of the file pppdump/pppdump.c of the
component pppdump. The manipulation of the argument spkt.buf/rpkt.buf
leads to improper validation of array index. The real existence of this
vulnerability is still doubted at the moment. The name of the patch is
a75fb7b198eed50d769c80c36629f38346882cbf. It is recommended to apply a
patch to fix this issue. VDB-216198 is the identifier assigned to this
vulnerability. NOTE: pppdump is not used in normal process of setting up
a PPP connection, is not installed setuid-root, and is not invoked
automatically in any scenario.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2022-4603

Patch from:
https://github.com/ppp-project/ppp/commit/a75fb7b198eed50d769c80c36629f38346882cbf

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20230129

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>