Revert "qemu: Fix CVE-2022-4144"
Issue: LIN1022-3321

This reverts commit 4cb3874abf4fdeb04337a48a14c765ba9b2269d4.

To fix the following build error
../qemu-6.2.0/hw/display/qxl.c: In function 'qxl_phys2virt':
../qemu-6.2.0/hw/display/qxl.c:1477:67: error: 'size' undeclared (first use in
this function); did you mean 'gsize'?
 1477 |         if (!qxl_get_check_slot_offset(qxl, pqxl, &slot, &offset, size)) {
      |                                                                   ^~~~
      |                                                                   gsize

The upstream maintainer has noticed this error and decided a week ago to revert
the patch introducing it.
https://lore.kernel.org/openembedded-core/CAGDS+nmtWaY7wR2YARRRv=kjUQd+39o9QL7v7XyBBfZWVoi=JA@mail.gmail.com/

But the revert has not been on either kirkstone or kirtstone-next. This is to
revert it locally and should be dropped once it's fixed upstream.

(LOCAL REV: NOT UPSTREAM) -- WRLinux specific

Signed-off-by: He Zhe <zhe.he@windriver.com>