zip: fix configure check by using _Static_assert
Issue: LIN1022-4529

It's incorrect to run a cross-compiled program on build machine
to check if some feature is available or not. As these two checks
in zip are basically just checking the size, we can use _Static_assert
and sizeof to do such check at compile time.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20230619

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>